requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
pillow>=8.0.0 